feat(node): Add back support for TypeScript 3.2
