+++ types/node/package.json
@@ -2,6 +2,11 @@
     "private": true,
     "types": "index",
     "typesVersions": {
+        ">=3.2.0-0": {
+            "*": [
+                "ts3.2/*"
+            ]
+        },
         ">=3.5.0-0": {
             "*": [
                 "ts3.5/*"
