+++ types/node/ts3.2/tslint.json
@@ -1,0 +1,10 @@
+{
+  "extends": "dtslint/dt.json",
+  "rules": {
+    "ban-types": false,
+    "unified-signatures": false,
+    "no-empty-interface": false,
+    "no-single-declare-module": false,
+    "strict-export-declare-modifiers": false // http2 needs this
+  }
+}
