+++ types/node/ts3.5/node-tests.ts
@@ -1,12 +1,13 @@
 // tslint:disable-next-line:no-bad-reference
-import '../node-tests';
+import '../ts3.2/node-tests';
 import '../assert';
 import '../buffer';
 import '../child_process';
 import '../cluster';
 import '../crypto';
 import '../dgram';
-import '../global';
+import '../ts3.2/fs';
+import '../ts3.2/global';
 import '../http';
 import '../http2';
 import '../net';
@@ -20,59 +21,13 @@ import '../repl';
 import '../stream';
 import '../tls';
 import '../tty';
-import '../util';
+import '../ts3.2/util';
 import '../worker_threads';
 import '../zlib';
 
-import { types, promisify } from 'util';
-import { BigIntStats, statSync, Stats } from 'fs';
-
-//////////////////////////////////////////////////////////
-/// Global Tests : https://nodejs.org/api/global.html  ///
-//////////////////////////////////////////////////////////
-{
-    const hrtimeBigint: bigint = process.hrtime.bigint();
-
-    process.allowedNodeEnvironmentFlags.has('asdf');
-}
-
-// Util Tests
-{
-    const value: BigInt64Array | BigUint64Array | number = [] as any;
-    if (types.isBigInt64Array(value)) {
-        // $ExpectType BigInt64Array
-        const b = value;
-    } else if (types.isBigUint64Array(value)) {
-        // $ExpectType BigUint64Array
-        const b = value;
-    } else {
-        // $ExpectType number
-        const b = value;
-    }
-
-    const arg1UnknownError: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: unknown, result: number) => void): void => { });
-    const arg1AnyError: (arg: string) => Promise<number> = promisify((arg: string, cb: (err: any, result: number) => void): void => { });
-}
-
-// FS Tests
-{
-    const bigStats: BigIntStats = statSync('.', { bigint: true });
-    const anyStats: Stats | BigIntStats = statSync('.', { bigint: Math.random() > 0.5 });
-}
-
-// Global Tests
-
-{
-    const a = Buffer.alloc(1000);
-    a.writeBigInt64BE(123n);
-    a.writeBigInt64LE(123n);
-    a.writeBigUInt64BE(123n);
-    a.writeBigUInt64LE(123n);
-    let b: bigint = a.readBigInt64BE(123);
-    b = a.readBigInt64LE(123);
-    b = a.readBigUInt64LE(123);
-    b = a.readBigUInt64BE(123);
-}
+/////////////////////////////////////////////////////
+/// WASI tests : https://nodejs.org/api/wasi.html ///
+/////////////////////////////////////////////////////
 
 import { WASI } from 'wasi';
